CONVERTPDF-main/requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
tabula-py>=2.9.0
fpdf>=1.7.2
